requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.2
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
django-fernet-encrypted-fields==0.3.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pycparser==2.22
PyJWT==1.7.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.4.0
